Fixing typecheck, minor analysis/harmonics fixes
