+++ src/modes/analysis/types.js
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Type definitions for Analysis mode
+ */
+
+/**
+ * Marker object for analysis mode
+ * @typedef {Object} AnalysisMarker
+ * @property {string} id - Unique marker identifier
+ * @property {string} color - Marker color
+ * @property {number} time - Time coordinate
+ * @property {number} freq - Frequency coordinate
+ */
