+++ src/modes/harmonics/HarmonicsMode.js
@@ -1,8 +1,9 @@
 import { BaseMode } from '../BaseMode.js'
 // SVG utilities removed - no display element
 import { updateHarmonicPanelContent, createHarmonicPanel } from '../../components/HarmonicPanel.js'
+import { showManualHarmonicModal } from './ManualHarmonicModal.js'
 import { notifyStateListeners } from '../../core/state.js'
-import { createLEDDisplay, createColorPicker, createFullFlexLayout, createFlexColumn, createFlexLayout } from '../../components/UIComponents.js'
+import { createColorPicker, createFullFlexLayout, createFlexColumn, createFlexLayout } from '../../components/UIComponents.js'
 
 /**
  * Harmonics mode implementation
@@ -30,10 +31,10 @@ export class HarmonicsMode extends BaseMode {
 
   /**
    * Handle mouse move events in harmonics mode
-   * @param {MouseEvent} event - Mouse event
-   * @param {Object} dataCoords - Data coordinates {freq, time}
+   * @param {MouseEvent} _event - Mouse event (unused in current implementation)
+   * @param {Object} _dataCoords - Data coordinates {freq, time} (unused in current implementation)
    */
-  handleMouseMove(event, dataCoords) {
+  handleMouseMove(_event, _dataCoords) {
     // Harmonics mode specific handling can be added here
   }
 
@@ -193,6 +194,7 @@ export class HarmonicsMode extends BaseMode {
       color = HarmonicsMode.harmonicColors[colorIndex]
     }
     
+    /** @type {HarmonicSet} */
     const harmonicSet = {
       id,
       color,
@@ -369,108 +371,7 @@ export class HarmonicsMode extends BaseMode {
    * Show manual harmonic modal dialog
    */
   showManualHarmonicModal() {
-    // Create modal overlay
-    const overlay = document.createElement('div')
-    overlay.className = 'gram-frame-modal-overlay'
-    
-    // Create modal dialog
-    const modal = document.createElement('div')
-    modal.className = 'gram-frame-modal'
-    
-    // Create modal content
-    modal.innerHTML = `
-      <div class="gram-frame-modal-header">
-        <h3>Add Manual Harmonics</h3>
-      </div>
-      <div class="gram-frame-modal-body">
-        <label for="harmonic-spacing-input">Harmonic spacing (Hz):</label>
-        <input type="number" id="harmonic-spacing-input" min="1.0" step="0.1" placeholder="Enter spacing in Hz">
-        <div class="gram-frame-modal-error" id="spacing-error" style="display: none; color: red; font-size: 12px; margin-top: 5px;">
-          Please enter a number ≥ 1.0
-        </div>
-      </div>
-      <div class="gram-frame-modal-footer">
-        <button class="gram-frame-modal-cancel" id="cancel-button">Cancel</button>
-        <button class="gram-frame-modal-add" id="add-button" disabled>Add</button>
-      </div>
-    `
-    
-    overlay.appendChild(modal)
-    document.body.appendChild(overlay)
-    
-    // Get modal elements
-    const spacingInput = /** @type {HTMLInputElement} */ (modal.querySelector('#harmonic-spacing-input'))
-    const errorDiv = /** @type {HTMLDivElement} */ (modal.querySelector('#spacing-error'))
-    const cancelButton = /** @type {HTMLButtonElement} */ (modal.querySelector('#cancel-button'))
-    const addButton = /** @type {HTMLButtonElement} */ (modal.querySelector('#add-button'))
-    
-    // Input validation
-    const validateInput = () => {
-      const value = parseFloat(spacingInput.value)
-      const isValid = !isNaN(value) && value >= 1.0
-      
-      if (spacingInput.value.trim() === '') {
-        // Empty input - hide error, disable button
-        errorDiv.style.display = 'none'
-        addButton.disabled = true
-      } else if (!isValid) {
-        // Invalid input - show error, disable button
-        errorDiv.style.display = 'block'
-        addButton.disabled = true
-      } else {
-        // Valid input - hide error, enable button
-        errorDiv.style.display = 'none'
-        addButton.disabled = false
-      }
-    }
-    
-    // Add input event listeners
-    spacingInput.addEventListener('input', validateInput)
-    spacingInput.addEventListener('keydown', (e) => {
-      if (e.key === 'Enter' && !addButton.disabled) {
-        addHarmonic()
-      } else if (e.key === 'Escape') {
-        closeModal()
-      }
-    })
-    
-    // Close modal function
-    const closeModal = () => {
-      document.body.removeChild(overlay)
-    }
-    
-    // Add harmonic function
-    const addHarmonic = () => {
-      const spacing = parseFloat(spacingInput.value)
-      if (!isNaN(spacing) && spacing >= 1.0) {
-        // Determine anchor time: use cursor position if available, otherwise midpoint
-        let anchorTime
-        if (this.state.cursorPosition) {
-          anchorTime = this.state.cursorPosition.time
-        } else {
-          // Use midpoint of time range
-          anchorTime = (this.state.config.timeMin + this.state.config.timeMax) / 2
-        }
-        
-        // Add the harmonic set
-        this.addHarmonicSet(anchorTime, spacing)
-        
-        // Close modal
-        closeModal()
-      }
-    }
-    
-    // Event listeners
-    cancelButton.addEventListener('click', closeModal)
-    addButton.addEventListener('click', addHarmonic)
-    overlay.addEventListener('click', (e) => {
-      if (e.target === overlay) {
-        closeModal()
-      }
-    })
-    
-    // Focus the input
-    spacingInput.focus()
+    showManualHarmonicModal(this, this.state, this.addHarmonicSet.bind(this))
   }
